# Agent-docs/05-api/openapi.yaml
openapi: 3.0.3
info:
  title: Jarvis Agent Platform API - v1.7.3 Studio Safety, Avatar Policy & Pricing
    Rules CRUD
  version: 1.7.3
  description: Jarvis Platform API Contract v1.7.1. Adds Agent Studio profile editor
    support and editable Pricing Packages CRUD in sandbox mode.
servers:
- url: /api/v1
security:
- bearerAuth: []
paths:
...
  /agents:
    get:
      tags:
      - Agent
      summary: List published agents
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentPublicSummary'
                  pageInfo:
                    $ref: '#/components/schemas/PageInfo'
                  requestId:
                    type: string
  /agents/{slug}:
    get:
      tags:
      - Agent
      summary: Get public agent detail
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Agent detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AgentPublicDetail'
                  requestId:
                    type: string
...
  /studio/agents:
    get:
      summary: List Studio agents
      tags:
      - Studio Agents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      summary: Create Studio agent draft
      tags:
      - Studio Agents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /studio/agents/{agentId}:
    get:
      tags:
      - Studio
      summary: Get studio agent
      parameters:
      - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Studio agent
...
  /studio/agents/{agentId}/publish:
    post:
      tags:
      - Studio
      summary: Publish agent version
      parameters:
      - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - version
              properties:
                version:
                  type: string
                releaseNote:
                  type: string
      responses:
        '200':
          description: Published
...
  /studio/model-profiles:
    get:
      summary: List Model Profiles
      tags:
      - Model Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      summary: Create Model Profile
      tags:
      - Model Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /studio/model-profiles/{id}:
    get:
      summary: Get Model Profile
      tags:
      - Model Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    patch:
      summary: Update Model Profile
      tags:
      - Model Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
    delete:
      summary: Delete Model Profile with Studio double-confirm UI
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: &id001
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                      id:
                        type: string
  /studio/model-profiles/{id}/test:
    post:
      summary: Test Model Profile
      tags:
      - Model Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /studio/voice-profiles:
    get:
      summary: List Voice Profiles
      tags:
      - Voice Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      summary: Create Voice Profile
      tags:
      - Voice Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /studio/voice-profiles/{id}:
    get:
      summary: Get Voice Profile
      tags:
      - Voice Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    patch:
      summary: Update Voice Profile
      tags:
      - Voice Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
    delete:
      summary: Delete Voice Profile with Studio double-confirm UI
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: *id001
  /studio/voice-profiles/{id}/test:
    post:
      summary: Test Voice Profile
      tags:
      - Voice Profiles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /billing/recharge-orders:
    post:
      summary: Create recharge order
      tags:
      - Billing
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
    get:
      summary: List current user visible recharge orders
      tags:
      - Billing
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /studio/users/{id}/tokens/adjust:
    post:
      summary: Admin manual token adjustment
      tags:
      - Billing
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /studio/agents/{id}:
    get:
      summary: Get Studio agent
      tags:
      - Studio Agents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    patch:
      summary: Update Studio agent
      tags:
      - Studio Agents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
    delete:
      summary: Delete Agent with Studio double-confirm UI
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: *id001
...
  /studio/billing/packages:
    get:
      summary: List editable recharge packages for Admin Studio
      tags:
      - Studio Billing
      responses:
        '200':
          description: OK
    post:
      summary: Create recharge package in sandbox
      tags:
      - Studio Billing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                amountRmb:
                  type: number
                agentTokens:
                  type: integer
                status:
                  type: string
                sortOrder:
                  type: integer
              required:
              - name
              - amountRmb
              - agentTokens
      responses:
        '200':
          description: OK
  /studio/billing/packages/{id}:
    patch:
      summary: Update recharge package in sandbox
      tags:
      - Studio Billing
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
    delete:
      summary: Delete recharge package in sandbox
      tags:
      - Studio Billing
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
  /studio/billing/packages/{id}/discount:
    post:
      summary: Apply quick bonus-token discount to a recharge package
      tags:
      - Studio Billing
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bonusPercent:
                  type: number
                reset:
                  type: boolean
      responses:
        '200':
          description: OK
  /studio/billing/pricing-rules:
    get:
      summary: List pricing rules for Studio CRUD
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items: &id002
                      type: object
                      properties:
                        id:
                          type: string
                        key:
                          type: string
                        label:
                          type: string
                        group:
                          type: string
                        valueType:
                          type: string
                        value: {}
                        unit:
                          type: string
                        description:
                          type: string
                        editable:
                          type: boolean
                        status:
                          type: string
                        sortOrder:
                          type: integer
                        createdAt:
                          type: string
                        updatedAt:
                          type: string
    post:
      summary: Create pricing rule
      requestBody:
        required: true
        content:
          application/json:
            schema: *id002
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: *id002
  /studio/billing/pricing-rules/{id}:
    patch:
      summary: Update pricing rule
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema: *id002
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: *id002
    delete:
      summary: Delete pricing rule with Studio double-confirm UI
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: *id001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Use `Device <device_token>`.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
        maximum: 100
...
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    AgentPublicSummary:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        avatarUrl:
          type: string
          nullable: true
        status:
          type: string
          enum:
          - PUBLISHED
        version:
          type: string
        capabilities:
          $ref: '#/components/schemas/AgentCapabilities'
    AgentPublicDetail:
      allOf:
      - $ref: '#/components/schemas/AgentPublicSummary'
      - type: object
        properties:
          social:
            $ref: '#/components/schemas/AgentSocialGallery'
          voice:
            type: object
            additionalProperties: true
    AgentSocialGallery:
      type: object
      properties:
        galleryImages:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              alt:
                type: string
                nullable: true
              sortOrder:
                type: integer
                nullable: true
        galleryVideos:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              posterUrl:
                type: string
                nullable: true
              title:
                type: string
                nullable: true
              sortOrder:
                type: integer
                nullable: true
    AgentCapabilities:
      type: object
      properties:
        chat:
          type: boolean
        voice:
          type: boolean
        tools:
          type: array
          items:
            type: string
...
    PageInfo:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
        hasNextPage:
          type: boolean
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data: {}
        message:
          type: string
        meta:
          type: object
        error:
          type: object
...